Revert change for interactive prompt to store cloud to pixie config file (#1964) (#1990)
Summary: Revert change for interactive prompt to store cloud to pixie
config file (#1964)

We've received feedback from users that is confusing to no longer have a
default cloud for the pixie cli. We will be restoring the previous
behavior in a follow up PR. With that new direction in mind, it no
longer makes sense to have the functionality added in #1964.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: None since this was a `git revert`

Signed-off-by: Dom Del Nano <ddelnano@gmail.com>
GitOrigin-RevId: 33244cd28d79058792a6cc06fe2637f491cd51a7