Add TLS protocol stitcher (#2058)
Summary: Add TLS protocol stitcher

This is meant to be a relatively bare bones stitcher at first. It will
be later extended but I wanted to get the minimal viable version in
first. My plan is to revisit this and other parser enhancements as an
excuse to "build in public" and video record hacking on Pixie's socket
tracer.

Relevant Issues: N/A

Type of change: /kind feature

Test Plan: Verified the following
- [x] New stitcher tests pass
- [x] Stitcher is functional with the rest of the TLS protocol changes
in place

![tls](https://github.com/user-attachments/assets/4c98dfae-0c9b-4c8a-a524-9ff3ffb3ec45)

Signed-off-by: Dom Del Nano <ddelnano@gmail.com>
GitOrigin-RevId: 1f6d18bfba7bd3cedf99e3d7eec5245b0e018f21