Add go 1.22 and 1.23 to socket tracer tests. Bump minor go versions (#2116)
Summary: Add go 1.22 and 1.23 to socket tracer tests. Bump minor go
versions

This PR adds Go 1.22 and 1.23 to the socket tracer test suite in
anticipation of removing Go 1.16 and 1.17 from our repo once #2108 is
merged. The plan is to follow this up with a change that removes Go 1.16
and 1.17 in addition to upgrading a portion of our Go dependencies. The
bulk of our dependencies still support Go 1.18, so this will allow us to
upgrade to the latest versions.

Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Test suite passes

---------

Signed-off-by: Dom Del Nano <ddelnano@gmail.com>
GitOrigin-RevId: 09ade3b0e8b255be582912d8d58ef390e0b90196