Replace `GUARDED_BY` with `ABSL_GUARDED_BY` in preparation for abseil upgrade (#2283)
Summary: Replace `GUARDED_BY` with `ABSL_GUARDED_BY` in preparation for
abseil upgrade

In order to upgrade to bazel 7, many of our dependencies (bazel repos)
need to be upgraded. In a branch that builds the pixie repo with bazel
7, this was a supporting change I needed to get the build working.
https://github.com/abseil/abseil-cpp/commit/ba7a9e2422cce2bc310c93dfb0b5997f69f7a291
is where abseil removed the versions of the macro pixie used previously.

Relevant Issues: #2282

Type of change: /kind cleanup

Test Plan: Build should succeed

---------

Signed-off-by: Dom Del Nano <ddelnano@gmail.com>
GitOrigin-RevId: 770c5091752326582bfb179892b180970b265515