Upgrade bazel to v6.5.0, upgrade protobuf, tensorflow and dependent deps (#2290)
Summary: Upgrade Bazel from 6.2.0 to 6.5.0 along with protobuf,
tensorflow, and related dependencies. This is a coordinated upgrade
since these dependencies have hard interdependencies that make
incremental upgrades difficult.

Key changes
- Upgrade Bazel from 6.2.0 to 6.5.0 (required by these new dependencies)
- Protobuf upgrade with compatibility patches for text format handling
and JavaScript generation
- TensorFlow upgrade with patches to disable GPU/LLVM/Python features.
The GPU disable patch will not be necessary as TensorFlow's fallback
logic will work with Bazel 7
- gRPC-web upgrade to 2.0.2 and replacement of the vendored
`protoc-gen-grpc-web` with a Bazel repository
- Flatbuffers patched to remove rules_js dependency (this upgrade is
required by TensorFlow v2.20.0)
- absl upgrade (causes transitive header changes)
- Regenerated TypeScript protobuf bindings for the UI

I attempted to split this into a smaller change in #2296. TensorFlow has
a tight dependency on protobuf, which prevented that attempt from
working. I'm open to other ideas on how to split this up, but so far
this seemed like the best balance with #2293 and #2297 to follow to
complete the Bazel 7 migration.

Relevant Issues: #2282

Type of change: /kind cleanup

Test Plan: Build passes

---------

Signed-off-by: Dom Del Nano <ddelnano@gmail.com>
GitOrigin-RevId: a507457029026e993360ab43f9480f811d35b5c8